+++ lib/vagrant/action/env.rb
@@ -1,0 +1,7 @@
+module Vagrant
+  class Action
+    module Env
+      autoload :Set, 'vagrant/action/env/set'
+    end
+  end
+end