+++ lib/vagrant/systems.rb
@@ -1,0 +1,8 @@
+# These can't be autoloaded because they have to register functionality
+# with Vagrant core.
+require 'vagrant/systems/base'
+require 'vagrant/systems/linux'
+require 'vagrant/systems/solaris'
+
+require 'vagrant/systems/debian'
+require 'vagrant/systems/gentoo'