+++ lib/vagrant/test_helpers.rb
@@ -31,8 +31,8 @@ module Vagrant
 
       str  = args.shift || ""
       File.open(path.to_s, "w") do |f|
+        f.puts "ENV['VAGRANT_HOME'] = '#{home_path}'"
         f.puts "Vagrant::Config.run do |config|"
-        f.puts "config.vagrant.home = '#{home_path}'"
         f.puts "config.vm.base_mac = 'foo' if !config.vm.base_mac"
         f.puts "config.vm.box = 'base'"
         f.puts str